moving StyledModal to common components folder
