+++ src/app/components/common/StyledModal/index.ts
